refactor(cosmic-swingset): Move chainStorage logic into a library
